Generalize ServiceMaintenance to operate on interator and move to shared crate for reuse
